+++ asyncpraw/models/reddit/wikipage.py
@@ -275,11 +275,6 @@ class WikiPage(RedditBase):
         self.__dict__.update(data)
         self._fetched = True
 
-    async def _fetch_data(self):
-        name, fields, params = self._fetch_info()
-        path = API_PATH[name].format(**fields)
-        return await self._reddit.request(method="GET", params=params, path=path)
-
     def _fetch_info(self):
         return (
             "wiki_page",